moved various MTV files under new mtv/checkout folder
