+++ client/app/queue/mtv/checkout/RemoveDecisionIssueModal.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import COPY from '../../../COPY';
-import Modal from '../../components/Modal';
+import COPY from '../../../../COPY';
+import Modal from '../../../components/Modal';
 
 export const RemoveDecisionIssueModal = ({ onCancel, onSubmit }) => {
   const buttons = [
