+++ client/app/queue/mtv/checkout/SubmitVacatedDecisionsView.jsx
@@ -2,8 +2,8 @@ import React, { useContext, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { MotionToVacateContext } from './MotionToVacateContext';
 import { useDispatch } from 'react-redux';
-import { editStagedAppeal, stageAppeal, setDecisionOptions } from '../QueueActions';
-import SubmitDecisionView from '../SubmitDecisionView';
+import { editStagedAppeal, stageAppeal, setDecisionOptions } from '../../QueueActions';
+import SubmitDecisionView from '../../SubmitDecisionView';
 import { useParams } from 'react-router';
 import DECISION_TYPES from '../../../constants/APPEAL_DECISION_TYPES';
